requirements.txt
httpx==0.27.2
python-dotenv==1.0.1
pydantic==2.9.2
orjson==3.10.7
